Move ASIO audio_device class into its own file
